Makefile simplifications
Checking for dynlink and correct file extensions manually is not
necessary as all of that information can come directly from `$(ocaml
-where)/Makefile.config`. It defines:

* `NATDYNLINK` which is true when dynlink is available
* `EXE` contains the platform specific extension for binaries
* `EXT_{LIB,OBJ}` contain platform specific lib/object extensions

Also, findlib installation is simplified to just use `$(wildcard ...)` to
select files to copy. wildcard is safe because it returns blank when the
targets aren't there.

Finally, there's no need to check for particularly binaries to install,
since we know what the name should be from `$(EXE)`.