Avoid full buffer copy in `map_lexeme`
As mentioned in #85 by @skcho, the code in `map_ident`/`map_string` was
changed to not copy the whole buffer and give a start/end point (though
this happened somewhat more on the accidental than deliberate side) thus
subtly changing the semantics of the data `f` was called on. This saves
copying but prevents `f` from peeking beyond the borders it was
assigned. Whether this breaks consumers depends on what they do with it.

Nevertheless, what this PR does is it applies the same logic to
`map_lexeme`, so while the semantics might have changed at least they
change in a consistent way.

Closes #85