Remove `validate_json` (#137)
It doesn't do much and made only sense in atdgen when we used to have a
`json` type. Now that the type doesn't exist it would require to be
renamed into `validate_t` but such functionality doesn't seem to make
sense to exist in Yojson, especially as it never does any validation to
begin with.