Use Command_unix following the deprecation of Command
In Core v0.15 `Command.run` is deprecated in favor of `Command_unix` to
be able to split Unix parts into their separate packages. This means the
code needs to adapt. Turns out v0.14 supports that already and since
this is a benchmarking binary there isn't that much value in keeping it
installable on large range of versions of Core.