Replace `core_bench` with `benchmark`
`core_bench` is nice but the amount of dependency issues and build
failures it causes via its `core_unix` dependency is draining a lot of
maintenance time trying to get it to work again so this replaces it with
a simpler library.

The output is a bit less fancy, but it should allow for similar
measurements.